Simplificado código de ProductScreen substituindo Stack e Box  por Flex
