+++ src/pages/ProductScreen.jsx
@@ -1,4 +1,4 @@
-import { Box, Stack } from "@chakra-ui/react"
+import { Box, Flex } from "@chakra-ui/react"
 import Navbar from "../components/Navbar"
 import ProductImage from "../components/product/ProductImage"
 import ProductTitle from "../components/product/ProductTitle"
@@ -12,23 +12,23 @@ function ProductScreen() {
         <>
 
         <Navbar></Navbar>
-        <Stack>
-            <Box display='flex' flexDirection='row'>
-                <ProductImage />
+       
+            <Flex>
+                <ProductImage src='https://i.discogs.com/yGmG6qztvTovGqYvQOefrsdlW5dLTIHdyrQVw2PgUfk/rs:fit/g:sm/q:90/h:554/w:554/czM6Ly9kaXNjb2dz/LWRhdGFiYXNlLWlt/YWdlcy9BLTY1MzY4/NTYtMTU3MjU0OTEx/OS02NjQzLmpwZWc.jpeg'/>
                 <Box ml='64px' mt='48px'>
-                    <ProductTitle></ProductTitle>
-                    <ProductArtist></ProductArtist>
-                    <ProductPrice></ProductPrice>
+                    <ProductTitle title='Ednaldo Pereira'></ProductTitle>
+                    <ProductArtist title='@ednaldopereira'></ProductArtist>
+                    <ProductPrice initialPrice={900} currentPrice={450} installments={10}></ProductPrice>
                     <ProductDescription></ProductDescription>
-                    <Box mt='48px' display='flex' flexDirection='row' gap='22px'>
+                    <Flex mt='48px' gap='22px'>
                         <ProductColor bg='#F6F099'></ProductColor>
                         <ProductColor bg='#F3985B'></ProductColor>
                         <ProductColor bg='#96D5FF'></ProductColor>
                         <ProductColor bg='#8CEAA6'></ProductColor>
-                    </Box>
+                    </Flex>
                 </Box>
-            </Box>
-        </Stack>
+            </Flex>
+        
 
         </>
     )
